HR-Agent-MatchMaking/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
opencv-python==4.10.0.84
Pillow==10.4.0
tensorflow==2.17.0
python-dotenv==1.0.1
pyyaml==6.0.1
kagglehub
matplotlib
seaborn
plotly
wordcloud
nltk
spacy
pymongo>=3.12,<4.0

# API dependencies
fastapi
uvicorn
PyPDF2

# NLP & Embedding dependencies
sentence-transformers
transformers
torch
codecarbon